fix: preserve contact data on re-resolve of OOBI
Re-resolving an OOBI wipes out all contact data stored for a contact, which is troublesome if you have stored an alias or any other metadata about a contact. This fixes that. The Organizer.update function uses a Python dict merge to merge the result of the OOBI resolution call with any existing data stored for the given contact. This is a good default.

Organizer.set() or Organizer.replace() can be used whenever overriding or wiping out prior data is desired.